Add files via upload
The GIS_Batch_Processing_demo folder contains the jupyter notebook which holds the code and the instructions. In that directory are also python scripts which are used in the notebook.

The IceFrontPositions.gdb folder contains the shapefiles fro the demonstration. Unzip it and make a note of the directory where these files are located.